+++ frontend/src/routes/_authenticated/annotation-campaign/$campaignID.phase.$phaseType/import-annotations.module.scss
